requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==1.4.*
python-dotenv
pydantic[email]
email-validator
